add foodItem to cart API

+++ src/main/java/com/example/EatExpress/model/Cart.java
@@ -22,7 +22,7 @@ public class Cart
         int id;
 
 
-        int cartTotal;
+        double cartTotal;
 
 
         @OneToOne
